Trying to extricate Material Design and added new audio player custom element
In an attempt to begin trying to extricate the use of MDC elements from the site (regretting falling for that meme rn tbh fam), I made a bunch of changes to JS source.  Site should load faster now since fewer JS files are being requested.  Still a long way to go to fully removing MDC, but this was a good step.

Also I've decided to try leverage custom elements, so created new mlp-audio-player element to take care of the audio playing element and the topic display.